+++ CHANGELOG.md
@@ -91,15 +91,18 @@
 5. Use class for defining rules
 6. Hook to ActiveModel without having to include `Bali::Authorizer`
 
-== Version 6.0.0rc2 (Apr 28, 2020)
+== Version 6.0.0rc2 (Apr 27, 2020)
 
 1. Hook into ActionView and ActionController
 2. Not storing the ruler class inside a `RULE_CLASS_MAP` to avoid caching it
 3. Not storing the role field mapper inside a `TRANSLATED_SUBTARGET_ROLES`
 
-== Version 6.0.0rc3 (Apr 29, 2020)
+== Version 6.0.0rc3 (May 01, 2020)
 
 1. Built-in `be_able_to` matcher for RSpec
 2. Renaming `Bali::Printer.pretty_print` to `Bali::Printer.printable`
 3. Uniform `can?` and `cant?` authorizization checker
 4. Adding `scope` block
+
+== Version 6.0.0rc4
+1. Bug fix for `can?` and `cant?` when passing 2 arguments.